# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/crafts": {
            "get": {
                "description": "获取所有技艺列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "crafts"
                ],
                "summary": "获取技艺列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "排序字段",
                        "name": "order_by",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建新的技艺（仅管理员可操作）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "crafts"
                ],
                "summary": "创建技艺",
                "parameters": [
                    {
                        "description": "技艺信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.Craft"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "创建成功",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误、技艺名称已存在、分类不存在或无效的难度等级",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/crafts/category/{category_id}": {
            "get": {
                "description": "根据分类ID获取该分类下的所有技艺",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "crafts"
                ],
                "summary": "获取分类下的技艺",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "分类ID",
                        "name": "category_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "无效的分类ID",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/crafts/difficulty/{level}": {
            "get": {
                "description": "根据难度等级获取技艺列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "crafts"
                ],
                "summary": "获取指定难度技艺",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "难度等级",
                        "name": "level",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "无效的难度等级",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/crafts/{id}": {
            "get": {
                "description": "根据ID获取技艺信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "crafts"
                ],
                "summary": "获取技艺",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "技艺ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "无效的技艺ID",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "技艺不存在",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新技艺信息（仅管理员可操作）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "crafts"
                ],
                "summary": "更新技艺",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "技艺ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "技艺信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.Craft"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "技艺不存在",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "409": {
                        "description": "技艺名称已存在",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "删除技艺（仅管理员可操作）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "crafts"
                ],
                "summary": "删除技艺",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "技艺ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "无效的技艺ID",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "技艺不存在",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/crafts/{id}/with-category": {
            "get": {
                "description": "根据ID获取技艺信息及其所属分类",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "crafts"
                ],
                "summary": "获取技艺及分类",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "技艺ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "无效的技艺ID",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "技艺不存在",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/backend_pkg_response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "backend_internal_models.Craft": {
            "type": "object",
            "properties": {
                "category": {
                    "$ref": "#/definitions/backend_internal_models.ICHCategory"
                },
                "categoryID": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "description": "工艺概况",
                    "type": "string"
                },
                "difficulty": {
                    "description": "技艺难度等级",
                    "type": "integer"
                },
                "history": {
                    "description": "历史沿革",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "regionFeatures": {
                    "description": "不同地域的表现形式差异",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "tools": {
                    "description": "所需核心工具列表",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "backend_internal_models.ICHCategory": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "iconURL": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "level": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "parentID": {
                    "type": "integer"
                },
                "regionCode": {
                    "description": "分类的地域特色标识",
                    "type": "string"
                },
                "sortOrder": {
                    "description": "排序权重",
                    "type": "integer"
                },
                "status": {
                    "description": "分类启用状态",
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "backend_pkg_response.ErrorField": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "backend_pkg_response.PageInfo": {
            "type": "object",
            "properties": {
                "page_num": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "backend_pkg_response.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "error": {
                    "$ref": "#/definitions/backend_pkg_response.ErrorField"
                },
                "page": {
                    "$ref": "#/definitions/backend_pkg_response.PageInfo"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        }
    }
}